Quiz routes cannot be accessed out of sequence
The quiz sequence was enforced only by which quiz link was rendered.
This update adds a check in the beforeModel hook of the quiz route,
checking the quiz ID agains previously submitted quizzes. If a user
attempts to access a quiz route directly out of order (by modifying the
URL, or using the browser navigaiton buttons) they are redirected back
to the quizzes list page.